get_started: move obs saver to utils; 1-5 test passed on isaaclab
